Refactor notification system to use context API
Replaced individual notification state and logic in components with a shared NotificationContext. Centralized notification logic in App.js for improved consistency and reusability across Login, Register, and ResetPassword components.
