+++ src/components/Login.js
@@ -1,8 +1,9 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { useState, useEffect } from 'react';
+import { useState, useContext } from 'react';
 import { Container, Form, InputGroup, Button } from 'react-bootstrap';
-import { FaEnvelope, FaLock, FaEye, FaEyeSlash, FaSpinner } from 'react-icons/fa';
+import { FaEnvelope, FaLock, FaEye, FaEyeSlash } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
+import { NotificationContext } from '../App';
 
 function Login() {
     const [showPassword, setShowPassword] = useState(false);
@@ -10,9 +11,8 @@ function Login() {
     const [password, setPassword] = useState('');
     const [isEmailValid, setIsEmailValid] = useState(null);
     const [isPasswordValid, setIsPasswordValid] = useState(null);
-    const [showModal, setShowModal] = useState(false);
-    const [currentMessage, setCurrentMessage] = useState('');
-    const [showText, setShowText] = useState(false);
+
+    const { showNotification } = useContext(NotificationContext); // Use the context
 
     const togglePasswordVisibility = () => {
         setShowPassword(!showPassword);
@@ -23,7 +23,7 @@ function Login() {
         const isValid = emailRegex.test(value);
         setIsEmailValid(isValid);
         if (!isValid && value) {
-            displayNotification('Invalid email address. Please include "@" and a valid domain (e.g., .com).');
+            showNotification('Invalid email address. Please include "@" and a valid domain (e.g., .com).');
         }
         return isValid;
     };
@@ -33,7 +33,7 @@ function Login() {
         const isValid = passwordRegex.test(value);
         setIsPasswordValid(isValid);
         if (!isValid && value) {
-            displayNotification('Password: 6+ chars, 1 letter, 1 number');
+            showNotification('Password: 6+ chars, 1 letter, 1 number');
         }
         return isValid;
     };
@@ -42,21 +42,7 @@ function Login() {
         e.preventDefault();
         if (isEmailValid && isPasswordValid) {
             console.log('Login Data:', { email, password });
-            displayNotification(`Logged in with: ${email}, ${password}`);
-        }
-    };
-
-    const displayNotification = (message) => {
-        if (showModal) {
-            setShowModal(false);
-            setShowText(false);
-            setTimeout(() => {
-                setCurrentMessage(message);
-                setShowModal(true);
-            }, 500);
-        } else {
-            setCurrentMessage(message);
-            setShowModal(true);
+            showNotification(`Logged in with: ${email}, ${password}`);
         }
     };
 
@@ -71,27 +57,6 @@ function Login() {
         setPassword(value);
         validatePassword(value);
     };
-
-    useEffect(() => {
-        if (showModal) {
-            const textTimer = setTimeout(() => {
-                setShowText(true);
-            }, 1000);
-
-            const closeTimer = setTimeout(() => {
-                setShowText(false);
-                setTimeout(() => {
-                    setShowModal(false);
-                    setCurrentMessage('');
-                }, 500);
-            }, 3000);
-
-            return () => {
-                clearTimeout(textTimer);
-                clearTimeout(closeTimer);
-            };
-        }
-    }, [showModal]);
 
     return (
         <Container className="d-flex justify-content-center align-items-center min-vh-100 position-relative">
@@ -167,69 +132,6 @@ function Login() {
                     </p>
                 </Form>
             </div>
-
-            {showModal && (
-                <div
-                    className="position-fixed"
-                    style={{
-                        top: '10px',
-                        left: '50%',
-                        transform: 'translateX(-50%)',
-                        backgroundColor: '#212529',
-                        color: '#fff',
-                        padding: '0 20px',
-                        height: '40px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)',
-                        zIndex: 1000,
-                        animation: 'openDynamicIsland 1s ease-in-out forwards, closeDynamicIsland 1s ease-in-out 3s forwards',
-                    }}
-                >
-                    <FaSpinner
-                        style={{
-                            position: 'absolute',
-                            animation: 'spin 1s linear infinite',
-                            opacity: showText ? 0 : 1,
-                            transition: 'opacity 0.3s ease',
-                        }}
-                    />
-                    <p
-                        style={{
-                            margin: 0,
-                            fontSize: '14px',
-                            opacity: showText ? 1 : 0,
-                            transition: 'opacity 0.3s ease',
-                            whiteSpace: 'nowrap',
-                            overflow: 'hidden',
-                            textOverflow: 'ellipsis',
-                            maxWidth: '260px',
-                        }}
-                    >
-                        {currentMessage}
-                    </p>
-                </div>
-            )}
-
-            <style>
-                {`
-                    @keyframes openDynamicIsland {
-                        0% { transform: translateX(-50%) translateY(-100%); width: 40px; border-radius: 50%; opacity: 0; }
-                        50% { transform: translateX(-50%) translateY(0); width: 40px; border-radius: 50%; opacity: 1; }
-                        100% { transform: translateX(-50%) translateY(0); width: 300px; border-radius: 20px; opacity: 1; }
-                    }
-                    @keyframes closeDynamicIsland {
-                        0% { transform: translateX(-50%) translateY(0); width: 300px; border-radius: 20px; opacity: 1; }
-                        50% { transform: translateX(-50%) translateY(0); width: 40px; border-radius: 50%; opacity: 1; }
-                        100% { transform: translateX(-50%) translateY(-100%); width: 40px; border-radius: 50%; opacity: 0; }
-                    }
-                    @keyframes spin {
-                        0% { transform: rotate(0deg); }
-                        100% { transform: rotate(360deg); }
-                    }
-                `}
-            </style>
         </Container>
     );
 }
