+++ src/components/ResetPassword.js
@@ -1,15 +1,12 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect, useRef, useContext } from 'react';
 import { Container, Form, InputGroup, Button } from 'react-bootstrap';
-import { FaEnvelope, FaLock, FaEye, FaEyeSlash, FaSpinner } from 'react-icons/fa';
-import { Link } from 'react-router-dom';
+import { FaEnvelope, FaLock, FaEye, FaEyeSlash } from 'react-icons/fa';
+import { NotificationContext } from '../App';
 
 function ResetPassword() {
     const [email, setEmail] = useState('');
     const [isEmailValid, setIsEmailValid] = useState(null);
-    const [showModal, setShowModal] = useState(false);
-    const [currentMessage, setCurrentMessage] = useState('');
-    const [showText, setShowText] = useState(false);
     const [step, setStep] = useState(1);
     const [code, setCode] = useState(['', '', '', '', '', '']);
     const [isCodeValid, setIsCodeValid] = useState(null);
@@ -23,12 +20,14 @@ function ResetPassword() {
     const [showResendButton, setShowResendButton] = useState(false);
     const inputRefs = useRef([]);
 
+    const { showNotification } = useContext(NotificationContext); // Use the context
+
     const validateEmail = (value) => {
         const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
         const isValid = emailRegex.test(value);
         setIsEmailValid(isValid);
         if (!isValid && value) {
-            displayNotification('Invalid email address. Please include "@" and a valid domain (e.g., .com).');
+            showNotification('Invalid email address. Please include "@" and a valid domain (e.g., .com).');
         }
         return isValid;
     };
@@ -37,7 +36,7 @@ function ResetPassword() {
         const isValid = codeArray.every((digit) => /^\d$/.test(digit));
         setIsCodeValid(isValid);
         if (!isValid && codeArray.some((digit) => digit !== '')) {
-            displayNotification('Code must contain exactly 6 digits.');
+            showNotification('Code must contain exactly 6 digits.');
         }
         return isValid;
     };
@@ -47,7 +46,7 @@ function ResetPassword() {
         const isValid = passwordRegex.test(value);
         setIsNewPasswordValid(isValid);
         if (!isValid && value) {
-            displayNotification('Password: 6+ chars, 1 letter, 1 number');
+            showNotification('Password: 6+ chars, 1 letter, 1 number');
         }
         return isValid;
     };
@@ -56,7 +55,7 @@ function ResetPassword() {
         const isValid = value === newPassword;
         setIsConfirmPasswordValid(isValid);
         if (!isValid && value) {
-            displayNotification('Passwords do not match.');
+            showNotification('Passwords do not match.');
         }
         return isValid;
     };
@@ -114,34 +113,34 @@ function ResetPassword() {
         if (step === 1) {
             if (isEmailValid) {
                 console.log('Server response: 200, Email:', email);
-                displayNotification(`Verification code sent to ${email}`);
+                showNotification(`Verification code sent to ${email}`);
                 setStep(2);
                 setTimer(60);
                 setShowResendButton(false);
             } else if (email) {
                 console.log('Server response: 400, Invalid email');
-                displayNotification('Invalid email address. Please try again.');
+                showNotification('Invalid email address. Please try again.');
             } else {
-                displayNotification('Please enter your email address.');
+                showNotification('Please enter your email address.');
             }
         } else if (step === 2) {
             if (isCodeValid) {
                 console.log('Code verified:', code.join(''));
-                displayNotification('Code verified successfully!');
+                showNotification('Code verified successfully!');
                 setStep(3);
             } else if (code.some((digit) => digit !== '')) {
-                displayNotification('Invalid code. Please try again.');
+                showNotification('Invalid code. Please try again.');
             } else {
-                displayNotification('Please enter the verification code.');
+                showNotification('Please enter the verification code.');
             }
         } else if (step === 3) {
             if (isNewPasswordValid && isConfirmPasswordValid) {
                 console.log('Password reset successfully, New Password:', newPassword);
-                displayNotification('Password reset successfully!');
+                showNotification('Password reset successfully!');
             } else if (!isNewPasswordValid) {
-                displayNotification('Password: 6+ chars, 1 letter, 1 number');
+                showNotification('Password: 6+ chars, 1 letter, 1 number');
             } else if (!isConfirmPasswordValid) {
-                displayNotification('Passwords do not match.');
+                showNotification('Passwords do not match.');
             }
         }
     };
@@ -154,45 +153,10 @@ function ResetPassword() {
 
     const handleResendCode = () => {
         console.log('Resending code to:', email);
-        displayNotification(`Verification code resent to ${email}`);
+        showNotification(`Verification code resent to ${email}`);
         setTimer(60);
         setShowResendButton(false);
     };
-
-    const displayNotification = (message) => {
-        if (showModal) {
-            setShowModal(false);
-            setShowText(false);
-            setTimeout(() => {
-                setCurrentMessage(message);
-                setShowModal(true);
-            }, 500);
-        } else {
-            setCurrentMessage(message);
-            setShowModal(true);
-        }
-    };
-
-    useEffect(() => {
-        if (showModal) {
-            const textTimer = setTimeout(() => {
-                setShowText(true);
-            }, 1000);
-
-            const closeTimer = setTimeout(() => {
-                setShowText(false);
-                setTimeout(() => {
-                    setShowModal(false);
-                    setCurrentMessage('');
-                }, 500);
-            }, 3000);
-
-            return () => {
-                clearTimeout(textTimer);
-                clearTimeout(closeTimer);
-            };
-        }
-    }, [showModal]);
 
     useEffect(() => {
         if (step === 2 && timer > 0 && !showResendButton) {
@@ -283,30 +247,6 @@ function ResetPassword() {
                     </div>
                 </div>
             </div>
-
-            {showModal && (
-                <div className="position-fixed" style={{ top: '10px', left: '50%', transform: 'translateX(-50%)', backgroundColor: '#212529', color: '#fff', padding: '0 20px', height: '40px', display: 'flex', alignItems: 'center', justifyContent: 'center', boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)', zIndex: 1000, animation: 'openDynamicIsland 1s ease-in-out forwards, closeDynamicIsland 1s ease-in-out 3s forwards' }}>
-                    <FaSpinner style={{ position: 'absolute', animation: 'spin 1s linear infinite', opacity: showText ? 0 : 1, transition: 'opacity 0.3s ease' }} />
-                    <p style={{ margin: 0, fontSize: '14px', opacity: showText ? 1 : 0, transition: 'opacity 0.3s ease', whiteSpace: 'nowrap', overflow: 'hidden', textOverflow: 'ellipsis', maxWidth: '260px' }}>{currentMessage}</p>
-                </div>
-            )}
-
-            <style>{`
-                @keyframes openDynamicIsland {
-                    0% { transform: translateX(-50%) translateY(-100%); width: 40px; border-radius: 50%; opacity: 0; }
-                    50% { transform: translateX(-50%) translateY(0); width: 40px; border-radius: 50%; opacity: 1; }
-                    100% { transform: translateX(-50%) translateY(0); width: 300px; border-radius: 20px; opacity: 1; }
-                }
-                @keyframes closeDynamicIsland {
-                    0% { transform: translateX(-50%) translateY(0); width: 300px; border-radius: 20px; opacity: 1; }
-                    50% { transform: translateX(-50%) translateY(0); width: 40px; border-radius: 50%; opacity: 1; }
-                    100% { transform: translateX(-50%) translateY(-100%); width: 40px; border-radius: 50%; opacity: 0; }
-                }
-                @keyframes spin {
-                    0% { transform: rotate(0deg); }
-                    100% { transform: rotate(360deg); }
-                }
-            `}</style>
         </Container>
     );
 }
